pyproject.toml
[project]
name = "volpe-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools>=1.75.1",
    "numpy>=2.3.5",
]
